+++ R/utils.R
@@ -1,7 +1,8 @@
 #' @importFrom stats reorder setNames
 #' @importFrom rlang :=
 #' @importFrom lifecycle badge
+#' @importFrom ggplot2 autoplot
 NULL
 
 # To avoid R CMD check notes about undefined global variables
-utils::globalVariables(c("x_names", "rel_imp", "object", "new_data", "nn_module", "fit_class"))
+utils::globalVariables(c("x_names", "rel_imp", "object", "new_data", "nn_module", "fit_class", "epoch", "loss", "set"))